refactor(canvas): colocate chat panel hooks
