+++ apps/canvas-workspace/harness/validate/validation.yaml
@@ -47,7 +47,7 @@ pathRules:
       - src/shared/conversation-runtime.ts
       - src/renderer/src/agent-chat/runtime/**
       - src/renderer/src/components/chat/hooks/useChatComposerStateKeyed.ts
-      - src/renderer/src/components/chat/ChatPanel.tsx
+      - src/renderer/src/components/chat/ChatPanel/index.tsx
       - src/renderer/src/components/chat/ChatPageBody/index.tsx
       - harness/mock-llm.mjs
       - harness/mock-llm.test.mjs
