+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPanel/hooks/__tests__/useComposerRequest.test.tsx
@@ -2,8 +2,8 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import type { ChatComposerRequest } from '../types';
-import { useComposerRequest } from './useComposerRequest';
+import type { ChatComposerRequest } from '../../../types';
+import { useComposerRequest } from '../useComposerRequest';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
