+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPanel/hooks/__tests__/useStartSkillChat.test.tsx
@@ -2,9 +2,9 @@
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import { AppShellProvider } from '../../shell/AppShellProvider';
-import { useStartSkillChat } from './useStartSkillChat';
+import { I18nProvider } from '../../../../../i18n';
+import { AppShellProvider } from '../../../../shell/AppShellProvider';
+import { useStartSkillChat } from '../useStartSkillChat';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
