+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPanel/hooks/useStartSkillChat.ts
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useRef } from 'react';
-import { useI18n } from '../../../i18n';
-import { useAppShell } from '../../shell/AppShellProvider';
-import type { SessionBackEntry } from '../SessionBackBar';
+import { useI18n } from '../../../../i18n';
+import { useAppShell } from '../../../shell/AppShellProvider';
+import type { SessionBackEntry } from '../../SessionBackBar';
 
 interface NewSessionResult {
   ok: boolean;
